Move MarshalXML tests into files for the business logic since that behavior should exist regardless of whether the go fix is used
